Move getChangeTrackingVersionCustom into JdbcCTService.
